// components/common/listItemContainer/listItemContainerStyle.scss.d.ts
export type Styles = {
  'buttonAction': string;
  'buttonActionDelete': string;
  'buttonActionEdit': string;
  'buttonsActionsContainer': string;
  'componentContainer': string;
  'componentContainerDefaultStyle': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;